% Generated by build script
\name{pgkb_drug}
\alias{pgkb_drug}
\title{Look up a drug/chemical in PharmGKB}
\usage{
pgkb_drug()
}
\arguments{
\item{name}{Drug name (e.g. "warfarin", "clopidogrel", "simvastatin")}
}
\value{tibble: id, name, smiles, inchi, pediatric, type}
\description{Look up a drug/chemical in PharmGKB}
