% Generated by build script
\name{pgkb_gene}
\alias{pgkb_gene}
\title{Look up a gene in PharmGKB}
\usage{
pgkb_gene()
}
\arguments{
\item{symbol}{Gene symbol (e.g. "CYP2D6", "VKORC1", "SLCO1B1")}
}
\value{tibble: id, symbol, name, chromosome, cpic_gene, pharm_var_gene}
\description{Look up a gene in PharmGKB}
