% Generated by build script
\name{pgkb_variant}
\alias{pgkb_variant}
\title{Look up a genetic variant in PharmGKB}
\usage{
pgkb_variant()
}
\arguments{
\item{rsid}{Variant rsID (e.g. "rs1065852", "rs4149056", "rs9923231")}
}
\value{tibble: id, symbol, name, change_classification,}
\description{Look up a genetic variant in PharmGKB}
